Use single share/task locked alert class
(cherry picked from commit 4d6af02cc1aa69bf9263f3044263337e8fc37191)
