--- src/middlewared/middlewared/alert/source/afp.py
@@ -1,5 +1,0 @@
-from middlewared.alert.common.sharing_tasks import ShareLockedAlertClass
-
-
-class AFPShareLockedAlertClass(ShareLockedAlertClass):
-    pass
