--- src/middlewared/middlewared/alert/source/cloud_sync.py
@@ -1,5 +1,0 @@
-from middlewared.alert.common.sharing_tasks import TaskLockedAlertClass
-
-
-class CloudSyncTaskLockedAlertClass(TaskLockedAlertClass):
-    pass
