+++ src/middlewared/middlewared/alert/source/nfs_bindaddr.py
@@ -1,5 +1,4 @@
 from middlewared.alert.base import AlertClass, SimpleOneShotAlertClass, AlertCategory, AlertLevel
-from middlewared.alert.common.sharing_tasks import ShareLockedAlertClass
 
 
 class NFSBindAddressAlertClass(AlertClass, SimpleOneShotAlertClass):
@@ -7,8 +6,3 @@ class NFSBindAddressAlertClass(AlertClass, SimpleOneShotAlertClass):
     level = AlertLevel.WARNING
     title = "NFS Services Could Not Bind to Specific IP Addresses, Using 0.0.0.0"
     text = "NFS services could not bind to specific IP addresses, using 0.0.0.0."
-
-
-class NFSShareLockedAlertClass(ShareLockedAlertClass):
-
-    text = '%(type)s share with "%(paths)s" is unavailable because it uses a locked dataset.'
