+++ src/middlewared/middlewared/etc_files/ganesha/ganesha.conf
@@ -75,7 +75,7 @@ EXPORT_DEFAULTS
                 'Skipping generation of %r NFS share as the underlying resource is locked', share['id']
             )
             middleware.call_sync(
-                'alert.oneshot_create', 'NFSShareLocked', {**share, 'paths': ', '.join(share['paths'])}
+                'alert.oneshot_create', 'ShareLocked', {**share, 'identifier': ', '.join(share['paths']), 'type': 'NFS'}
             )
             continue
