+++ src/middlewared/middlewared/etc_files/local/apache24/Includes/webdav.conf
@@ -92,7 +92,9 @@ Listen ${webdav_config['tcpport']}
 			middleware.logger.debug(
 			    'Skipping generation of %r webdav share as underlying resource is locked', share['name']
 			)
-			middleware.call_sync('alert.oneshot_create', 'WebDAVShareLocked', share)
+			middleware.call_sync(
+			    'alert.oneshot_create', 'ShareLocked', {**share, 'identifier': share['name'], 'type': 'WebDAV'}
+			)
 			continue
 	%>\
 		Alias /${share['name']} "${share['path']}"
