+++ src/middlewared/middlewared/etc_files/local/rsyncd.conf
@@ -12,7 +12,9 @@ ${rsyncd_config['auxiliary']}
 % for mod in mods:
 <%
     if mod['locked']:
-        middleware.call_sync('alert.oneshot_create', 'RsyncModuleLocked', mod)
+        middleware.call_sync(
+            'alert.oneshot_create', 'ShareLocked', {**mod, 'identifier': mod['name'], 'type': 'Rsync Module'}
+        )
         continue
 %>\
 [${mod['name']}]
