+++ src/middlewared/middlewared/etc_files/local/smb4_share.conf
@@ -52,7 +52,9 @@
                     middleware.logger.debug(
                         "Skipping generation of %r share as the underlying resource is locked", share["name"]
                     )
-                    middleware.call_sync("alert.oneshot_create", "SMBShareLocked", share)
+                    middleware.call_sync(
+                        "alert.oneshot_create", "ShareLocked", {**share, "identifier": share["name"], "type": "SMB"}
+                    )
                     continue
                 if share["home"]:
                     share["name"] = "homes"
