+++ src/middlewared/middlewared/plugins/iscsi.py
@@ -474,8 +474,6 @@ class iSCSITargetExtentModel(sa.Model):
 
 class iSCSITargetExtentService(SharingService):
 
-    locked_alert_class = 'ISCSIExtentShareLocked'
-
     class Config:
         namespace = 'iscsi.extent'
         datastore = 'services.iscsitargetextent'
