+++ src/middlewared/middlewared/plugins/nfs.py
@@ -249,7 +249,6 @@ class NFSShareModel(sa.Model):
 
 class SharingNFSService(SharingService):
 
-    locked_alert_class = 'NFSShareLocked'
     path_field = 'paths'
 
     class Config:
